Moved unit tests into separate files. Added extra test.
