+++ README.md
@@ -16,14 +16,14 @@ The full list of checks managed by this package are provided below:
 
 ### Single file tests
 
-```
-temporal_subset must contain two components in the form "start"-"end", separated by "-"
-The structure of "start" and "end" must be of the form yyyy[MM[dd[hh[mm]]]]
-"start" and "end" must have the same the structure
-variable_table = "fx": time independent data will not have a temporal element
-variable_table = "yr": yearly data of the form: yyyy
-variable_table = "mon": monthly data of the form: yyyyMM
-variable_table = "monClim": monthly climatology data of the form: yyyyMM
+The single file tests are:
+ - The temporal subset must contain two components in the form "start"-"end", separated by "-"
+ - The structure of "start" and "end" must be of the form yyyy[MM[dd[hh[mm]]]]
+ - "start" and "end" must have the same the structure
+ - variable_table = "fx": time independent data will not have a temporal element
+ - variable_table = "yr": yearly data of the form: yyyy
+ - variable_table = "mon": monthly data of the form: yyyyMM
+ - variable_table = "monClim": monthly climatology data of the form: yyyyMM
 variable_table = "day": daily data of the form: yyyyMMdd
 variable_table = "6hr": 6 hourly data of the form: yyyyMMddhh
 variable_table = "3hr": 3 hourly data of the form: yyyyMMddhhmm
