+++ requirements.txt
@@ -1,5 +1,6 @@
 netCDF4
 arrow
+cf_units
 
 ###################
 # For testing:
